requirements.txt
numpy>=1.20.0
torch>=1.13.0
gymnasium>=0.28.1
matplotlib>=3.5.0
PyYAML>=6.0
tqdm>=4.62.0
tensorboard>=2.10.0
wandb>=0.13.0
opencv-python>=4.6.0.66
pillow>=9.2.0
imageio>=2.22.0
